+++ lib/swagger_docs_generator/parser/models/active_record.rb
@@ -5,6 +5,22 @@ module SwaggerDocsGenerator
   # # Parse Active Record model
   class ModelActiveRecord
     NAME = :active_record
+    TYPES = {
+      'binary' =>       { type: 'string', format: 'binary' },
+      'boolean' =>      { type: 'boolean' },
+      'date' =>         { type: 'string', format: 'date' },
+      'datetime' =>     { type: 'string', format: 'date-time' },
+      'decimal' =>      { type: 'number', format: 'double' },
+      'float' =>        { type: 'number', format: 'float' },
+      'integer' =>      { type: 'integer', format: 'int32' },
+      'bigint' =>       { type: 'integer', format: 'int64' },
+      'primary_key' =>  { type: 'integer', format: 'int32' },
+      'references' =>   { type: 'integer', format: 'int32' },
+      'string' =>       { type: 'string' },
+      'text' =>         { type: 'string' },
+      'time' =>         { type: 'string', format: 'date-time' },
+      'timestamp' =>    { type: 'string', format: 'date-time' }
+    }.freeze
 
     def initialize(model)
       @model = eval(model)
@@ -21,7 +37,9 @@ module SwaggerDocsGenerator
     private
 
     def attribute_propertie(name)
-      { name => { type: @model.columns_hash[name].type } }
+      {
+        name => TYPES[@model.columns_hash[name].type.to_s]
+      }
     end
   end
 end