+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.2.0] - 2017-01-06
+### [Fix]
+- Fix parameters body/other type
+- Fix definitions type/format
+
 ### [Changed]
 - Update responses
 - Update parameters